+++ src/user/users.module.ts
@@ -2,6 +2,7 @@ import { Module } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { UsersController } from './users.controller';
 import { PrismaService } from '@app/prisma/prisma.service';
+import { AuthModule } from '@app/auth/auth.module';
 
 @Module({
   controllers: [UsersController],
